pyproject.toml
[tool.poetry]
name = "patterned-api"
version = "0.1.0"
description = ""
authors = ["Daniel <gaievskyi@proton.me> Gaievskyi"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.11.3"
fastapi = "^0.54.1"
requests = "^2.23.0"
pydantic = "^1.4"
sqlalchemy = "^1.3.16"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
tenacity = "^6.1.0"
email-validator = "^1.0.5"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"